fix: refactor websockets module structure
